dc1/src/requirements.txt
numpy==1.26.4
pandas==2.2.0
Pillow==10.2.0
torch==2.1.2
torchvision==0.16.2
